feat: improve app installation
BREAKING CHANGE:
The third parameter `packageUri` is dropped from `db.app.install`.
If your code contains a call with client, xarName and packageUri this fails,
because the packageUri is now interpreted as the customPackageUrl.

Unpack a XAR in existdb to extract its package URI and use this info
to remove a previously installed version of this package.
This method is less error prone than to rely on the user to provide the
correct package URI themselves.

Since db.app.upload already hard-codes the collection
where packages are uploaded to (`/db/system/repo`), this
can safely assumed in `db.app.install`
when referencing the XAR by its name.

The improvements in the prepared query `install-package.xq`
was influenced by
https://github.com/eXist-db/atom-editor-support/blob/4715ec045bd6c305ca93003215714a494bad3fa3/packages.xql