ci: mark latest as experimental
Apparently, collection listings have a XML declaration, which was not
the case up until 7.0.0-SNAPSHOT. This being the new default would not
be a problem by itself but since serialization parameter
"omit-xml-declaration" is not respected this is a bug upstream and not
something that should be addressed in node-exist itself.

This results in two tests to fail with the current 7.0.0-SNAPSHOT of
exist-db and is the reason why this version is marked as experimental
and is allowed to fail.