feat(connection): align connection options
Both REST and XML-RPC now have the same set of default options.
The boolean option `secure` can still be used
to control XMLRPC client, but `protocol` is the preferred way.
This allows to use the same set of options for both client constructors.

Add test to ensure legacy option is still working.