fix(connection): misleading warning message
When connecting to a remote DB over an encrypted channel users might be
greeted with a warning:
"Connecting to remote DB allowing invalid certificate."
This is misleading in cases where `rejectUnauthorized` is not explicitly
set.

The added tests will connect to exist-db.org over https.
They only check the connection is established and rejected as
unauthorized.